Move parameter functions to methods.py
