+++ classes.py
@@ -184,189 +184,6 @@ class Class:
 
         return changed
     
-    #METHODS Parameters
-    def add_parameter(self, method_name, param_type, param_name, parameters=None):
-        """
-        Adds a parameter to a method.
-
-        Args:
-            method_name (str): The name of the method to which the parameter will be added.
-            param_type (str): The type of the parameter to be added.
-            param_name (str): The name of the parameter to be added.
-            parameters (list, optional): A list of parameters to match the method signature. Defaults to None.
-
-        Returns:
-            bool: True if the parameter was successfully added, False otherwise.
-        """
-        for method in self.methods:
-            if method.name == method_name:
-                should_add = False
-
-                # Case 1: parameters is None - add to all methods with this name
-                if parameters is None:
-                    should_add = True
-
-                # Case 2: empty parameters list - only add to no-parameter method
-                elif len(parameters) == 0 and len(method.parameters) == 0:
-                    should_add = True
-
-                # Case 3: specific parameters - add to matching method
-                elif parameters and method.matches_signature(parameters):
-                    should_add = True
-
-                if should_add:
-                    new_param = Parameter(param_name, param_type)
-                    for param in method.parameters:
-                        if param.getName() == param_name or param.getType() == param_type:
-                            return False
-                    method.parameters.append(new_param)
-                    return True
-        return False
-
-    def remove_parameter(self, method_name, param_name, parameters=None):
-        """
-        Removes a parameter from a method.
-
-        Args:
-            method_name (str): The name of the method from which the parameter will be removed.
-            param_name (str): The name of the parameter to be removed.
-            parameters (list, optional): A list of parameters to match the method signature. Defaults to None.
-
-        Returns:
-            bool: True if the parameter was successfully removed, False otherwise.
-        """
-        for method in self.methods:
-            if method.name == method_name:
-                should_remove = False
-
-                # Case 1: parameters is None - remove from all methods with this name
-                if parameters is None:
-                    should_remove = True
-
-                # Case 2: empty parameters list - only remove from no-parameter method
-                elif len(parameters) == 0 and len(method.parameters) == 0:
-                    should_remove = True
-
-                # Case 3: specific parameters - remove from matching method
-                elif parameters and method.matches_signature(parameters):
-                    should_remove = True
-
-                if should_remove:
-                    for param in method.parameters:
-                        if param.getName() == param_name:
-                            method.parameters.remove(param)
-                            return True
-        return False
-
-    def change_parameter(self, method_name, old_param_name, new_param_name, param_type, parameters=None):
-        """
-        Changes the name and type of a parameter in a method.
-
-        Args:
-            method_name (str): The name of the method containing the parameter to be changed.
-            old_param_name (str): The current name of the parameter to be changed.
-            new_param_name (str): The new name of the parameter.
-            param_type (str): The new type of the parameter.
-            parameters (list, optional): A list of parameters to match the method signature. Defaults to None.
-
-        Returns:
-            bool: True if the parameter was successfully changed, False otherwise.
-        """
-        for method in self.methods:
-            if method.name == method_name:
-                should_change = False
-
-                # Case 1: parameters is None - change in all methods with this name
-                if parameters is None:
-                    should_change = True
-
-                # Case 2: empty parameters list - only change in no-parameter method
-                elif len(parameters) == 0 and len(method.parameters) == 0:
-                    should_change = True
-
-                # Case 3: specific parameters - change in matching method
-                elif parameters and method.matches_signature(parameters):
-                    should_change = True
-
-                if should_change:
-                    for param in method.parameters:
-                        if param.getName() == old_param_name:
-                            param.setName(new_param_name)
-                            param.setType(param_type)
-                            return True
-        return False
-
-    def change_all_parameters(self, method_name, param_names, param_types, parameters=None):
-        """
-        Changes all parameters of a method.
-
-        Args:
-            method_name (str): The name of the method whose parameters will be changed.
-            param_names (list): A list of new parameter names.
-            param_types (list): A list of new parameter types.
-            parameters (list, optional): A list of parameters to match the method signature. Defaults to None.
-
-        Returns:
-            bool: True if the parameters were successfully changed, False otherwise.
-        """
-        for method in self.methods:
-            if method.name == method_name:
-                should_change = False
-
-                # Case 1: parameters is None - change in all methods with this name
-                if parameters is None:
-                    should_change = True
-
-                # Case 2: empty parameters list - only change in no-parameter method
-                elif len(parameters) == 0 and len(method.parameters) == 0:
-                    should_change = True
-
-                # Case 3: specific parameters - change in matching method
-                elif parameters and method.matches_signature(parameters):
-                    should_change = True
-
-                if should_change:
-                    if len(param_names) != len(param_types):
-                        return False
-                    method.parameters = [Parameter(name, type_) for name, type_ in zip(param_names, param_types)]
-                    return True
-        return False
-
-    def change_parameter_type(self, method_name, param_name, new_type, parameters=None):
-        """
-        Changes the type of a parameter in a method.
-
-        Args:
-            method_name (str): The name of the method containing the parameter to be changed.
-            param_name (str): The name of the parameter whose type will be changed.
-            new_type (str): The new type of the parameter.
-            parameters (list, optional): A list of parameters to match the method signature. Defaults to None.
-
-        Returns:
-            bool: True if the parameter type was successfully changed, False otherwise.
-        """
-        for method in self.methods:
-            if method.name == method_name:
-                should_change = False
-
-                # Case 1: parameters is None - change in all methods with this name
-                if parameters is None:
-                    should_change = True
-
-                # Case 2: empty parameters list - only change in no-parameter method
-                elif len(parameters) == 0 and len(method.parameters) == 0:
-                    should_change = True
-
-                # Case 3: specific parameters - change in matching method
-                elif parameters and method.matches_signature(parameters):
-                    should_change = True
-
-                if should_change:
-                    for param in method.parameters:
-                        if param.getName() == param_name:
-                            param.setType(new_type)
-                            return True
-        return False
